Updated version. Fixed load empty comments bug.

+++ Bash.Common/Data/BashClient.cs
@@ -87,19 +87,21 @@ namespace Bash.Common.Data
                 var encodedData = await ReadEncodedContentAsync(response);
                 var data = JsonConvert.DeserializeObject<BashComments>(encodedData);
                 
-                // trim to 50 items
-                if (data.Comments.Count > 50)
+                if (data != null)
                 {
-                    data.Comments.RemoveRange(50, data.Comments.Count - 50);
-                }
+                    // trim to 50 items
+                    if (data.Comments.Count > 50)
+                    {
+                        data.Comments.RemoveRange(50, data.Comments.Count - 50);
+                    }
 
-                for (int i = 0; i < data.Comments.Count; ++i)
-                {
-                    data.Comments[i].IndexPosition = i;
+                    for (int i = 0; i < data.Comments.Count; ++i)
+                    {
+                        data.Comments[i].IndexPosition = i;
+                    }
                 }
                 return data;
             }
-
             return null;
         }
